Add gulp-imagemin dependencies
Coming back to this project, these were for some reason required by
imagemin but not installed automatically as dependencies of it.

I hope to remove these in the future, I'll need to take a look at
imagemin another time to see why this is. I might need to find an
alternative if I can't remove them.