Remove excess content for the redesign
I want to start with a fresh slate, the bare minimum I need to get by. I
need to focus on content before I style it up.

There was too much wasted space on the previous design. At least in terms
of the masthead. A full-height background is great, but I wasn't utilizing
it properly.