Simplify the main stylesheet
This removes all the extra newlines and the Table of Contents at the top.

The ToC was tedious to maintain, the file is much easier to navigate
without it. It could be argued that the ToC would make it a lot easier
for newcomers to find their way around, but I don't want to maintain it
anymore.