Fix background not staying in place on mobile
It wasn't staying fixed, it was more like the background was following
after a delay. I've opted to simply apply .background onto html.

Inuit's base.page Sass file applies a background-color to the html
element, which appears to be the culprit. When that style is commented
out, everything works fine on mobile.