Retain bottom margins on lists
The left margin is really the only one we care about. We have other partials
removing all the other padding and margins.

We need to remove the left margin so the list stays lined up properly, but
we have to keep the bottom margin so as not to knock out vertical rhythm