Refactor how paths are handled.
All the commonly accessed source directories have been edited so we don't
need to prepend the source dir to all of them. We have a handle function
for that now.

The source and build dirs are also defined as constants now. This was
mostly done for the source dir so that we could access it outside of the
`paths` object, and the build dir was moved up too because it doesn't
really belong with those paths.