Just use one favicon.
There was actually no point in having the favicon.png file, that was just
done out of preference for PNG, and also so it could be grouped with the
other images.

The reason for having a favicon.ico in the root of the website is so web
browsers can display the favicon even on non-html content, like an image or
a PDF where you can't specify a <link> to the file.

http://stackoverflow.com/a/21359390