Move project images into their own folders
This also removes the 'hd' folder, that had the original project screenshots before compression.

This is likely a bad practice, but all images are uncompressed at the moment. This could potentially lead to performance implications, but for simplicity we're leaving them this way for right now.

Page load time shouldn't take too much of a hit from this, but if it does we can look into compressing all of the images before publishing.