Update to the latest version of inuitcss
We're installing through NPM and adding the inuitcss install to includePaths, so that we can acess it directly when importing.

This means we're no longer using Bower, so bower.json and .bowerrc can both be removed, along with the ignored bower_components folder. The root of the project is a lot cleaner, and now we only have to worry about a single package manager.