Remove the use of aliases
This was done to give us shorter variable names for things like the font-size mixin and spacing units. It doesn't really make much of a difference, and adds an unecessary layer of abstraction.

The default inuit variables are fine, we don't need a file dedicated to overriding them with shorter variants.