Begin the journey into Webpack
Currently you need to run webpack and then gulp. `npm start` has been configured to run them sequentially, so you can use that command for developing right now.

I'm still in the process of understanding Webpack and how to structure code around it. So far we're focusing on JavaScript but Sass, images, and (hopefully) Pug will be coming soon.

Attempts were made to get Pug compiling up and running, but they ultimately failed. We may have to move all our Pug templates to static HTML for the time being. This project will eventually be converted to React, but that is still a ways out.