Move the favicon into the images folder
It was in the root of the project so it would be compiled to the root with our Gulp setup.

This was so web browsers could automatically pull it in if they couldn't recognize the shortcut icon link tag, for whatever reason.

It's compiled to the root regardless now because of Webpack, so we can tuck it away with all the other images.