Consolidate Photo Components into one file + refactor photo creation
This should have been two separate commits, but these changed have been intermingled to the point where it would be difficult to split them. To save time we're commiting them together.

First set of changes consolidates all Photo related Components into a single file. It was getting a bit messy having them split up, and they're small enough where putting them in one file is completely fine and easy to read through.

The second set refactors how we define our photos. We now have a <Photo> Component which is used for creating everything except the caption (That's still handled by <PhotoCaption>, passed in as a child). This cleans up our page Components a lot, as we don't need to define a similar structure of Components repeatedly.