Remove images going from grayscale to color
Maintaining the plain JavaScript for this feature with the React code was a bit messy.

We should be able to implement this feature again purely in React, but for the time being we're going to remove it. While unconfirmed, this feature could also be jarring to new users. Images going from grayscale to color as you scroll past them isn't expected behavior, and could leave users confused about how it happened.