Use a single wrapper to contain everything
Instead of sectioning off our content, it all exists under a single wrapper now.

Idealy with a single-page website, you'd want IDs on your elements so the user can bookmark sections of the page and navigate using ID-based links.

We're going to be splitting content across multiple pages, and the portfolio is going to be setup in such a way where this kind of bookmarking won't be needed.