Merge routes with the router
This is how routes were initially setup, and were then moves to the `routes/` folder. Recently having the 'index.jsx' file inside routes/ has been getting annoying, as it's not technically a route but it's still mixed it.

Merging all of our routes back with the router means everything under routes/ is, in fact, a route.