Make the Webpack config a bit more readable
Newline between each object, and instead of 'jsx?' we're explicitly matching 'js' and 'jsx'.

The latter is just so it's easily apparent as to what we're doing. Those that aren't too familiar with regular expressions might be confused about it, whereas (js|jsx) is more intuitively understood as an 'or' operation.