Update the side-by-side comparisons
This is going to need to be revised in the future, as we currently have unwanted margin applied on the sides of the comparisons.

This is hard to fix, because we have to wrap all of our content for the project in a column so we can easily resize it.

Neat's coloumns don't play well when nested inside each other, and finding a fix has proven troublesome.